requirements.txt
python-dotenv

Flask
Flask-Bcrypt
PyJWT

alembic
psycopg2

beautifulsoup4
undetected-chromedriver
selenium

gnews
